Changed Scene * Renamed ShowProfile to ShowProfileScene * Added TokenID in editProfileScene
