+++ project/app/src/main/java/com/nexte/nexte/EditProfileScene/EditProfilePresenter.kt
@@ -8,15 +8,15 @@ interface EditProfilePresentationLogic{
     fun presentEditProfile(response: EditProfileModel.Response)
 }
 class EditProfilePresenter: EditProfilePresentationLogic{
-    var view: EditProfileDisplayLogic?= null
+    var view: EditProfileDisplayLogic? = null
 
     override fun presentEditProfile(response: EditProfileModel.Response){
         var message: String= ""
         if(response.sex == 'U'){
-            message= "Não foi possível editar o usuário"
+            message = "Não foi possível editar o usuário"
         }
         else {
-            message= "Recuperei o usuário"
+            message = "Recuperei o usuário"
         }
         var viewModel: EditProfileModel.ViewModel = EditProfileModel.ViewModel(message)
